refactor: Break main.js into modules
